selection items with pagination
